Refactoring and added RoverManagerTests
