+++ src/main/java/com/samuellohjw/marsroverapp/model/Rover.java
@@ -1,5 +1,7 @@
 package com.samuellohjw.marsroverapp.model;
 
+import java.util.Objects;
+
 public class Rover {
     private Position position;
 
@@ -11,10 +13,12 @@ public class Rover {
         return position;
     }
 
-    public void setPosition(Position position) {
-        this.position = position;
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Rover rover)) return false;
+        return Objects.equals(position, rover.position);
     }
-
 
 }
